test(2.0): rename test files and complete test case
